Reactivate the Maven 'clean' target.
This issue:

  https://jira.sakaiproject.org/browse/SAKIII-4367

saw 'mvn clean' get disabled.  This causes problems for the JavaScript/CSS hashing, since the output from one build ends up getting pulled in as input for the next.  Pages were rewritten to reference URLS like:

  /devwidgets/personinfo/css/personinfo-1db0f3e8daab32ed907b87f4b53ab7c4.css

but the file that ended up in the uxloader jar was:

  SLING-INF/content/devwidgets/personinfo/css/personinfo-1db0f3e8daab32ed907b87f4b53ab7c4-1db0f3e8daab32ed907b87f4b53ab7c4.css

(and I guess each run of mvn -Psakai-release would stick on another hash)